adding IAM requirements
As the app-dev profile includes both cloudwatch and CA components we need to accompany the demo with the requisite permissions to access the AWS API's. Im not sure if Kubernetes-system is the right place (although it is the right namespace and has associated components). This could live in a separate directory for clarity but either way we need to push this config so the Autoscaler doesnt hang and the cloudwatch logs can be pushed as the demo suggests.